temporary disable finalized. logs are now in txt format.

+++ app/src/main/java/balti/migratehelper/UninstallService.java
@@ -1,14 +1,15 @@
 package balti.migratehelper;
 
+import android.app.Notification;
 import android.app.NotificationManager;
 import android.app.Service;
 import android.content.ComponentName;
-import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.os.IBinder;
 import android.support.annotation.Nullable;
+import android.support.v4.app.NotificationCompat;
 import android.widget.Toast;
 
 import java.io.BufferedWriter;
@@ -16,19 +17,30 @@ import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.util.Objects;
 
-import static balti.migratehelper.AppSelector.TEMP_DIR_NAME;
+import static balti.migratehelper.CommonTools.TEMP_DIR_NAME;
+import static balti.migratehelper.Listener.PROGRESS_CHANNEL;
 
 public class UninstallService extends Service {
+
+    int UNINSTALL_START_ID = 234;
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
 
+        Notification uninstallNotif = new NotificationCompat.Builder(this, PROGRESS_CHANNEL)
+                .setSmallIcon(R.drawable.ic_notification_icon)
+                .setContentText(getString(R.string.finishing))
+                .build();
+
+        startForeground(UNINSTALL_START_ID, uninstallNotif);
 
         int dpiValue = intent.getIntExtra("dpiValue", 0);
+        boolean doReboot = intent.getBooleanExtra("doReboot", false);
+        boolean doUninstall = intent.getBooleanExtra("doUninstall", true);
         try {
 
             ((NotificationManager) Objects.requireNonNull(getSystemService(NOTIFICATION_SERVICE))).cancelAll();
-            uninstall(this, dpiValue);
+            finishTasks(dpiValue, doUninstall, doReboot);
 
         } catch (IOException | InterruptedException e) {
             e.printStackTrace();
@@ -38,9 +50,9 @@ public class UninstallService extends Service {
         return super.onStartCommand(intent, flags, startId);
     }
 
-    void uninstall(Context context, int dpiValue) throws IOException, InterruptedException {
+    void finishTasks(int dpiValue, boolean doUninstall, boolean doReboot) throws IOException, InterruptedException {
 
-        String sourceDir = context.getApplicationInfo().sourceDir;
+        String sourceDir = getApplicationInfo().sourceDir;
 
         Process fullProcess = Runtime.getRuntime().exec("su");
         BufferedWriter writer = new BufferedWriter(new OutputStreamWriter(fullProcess.getOutputStream()));
@@ -49,26 +61,29 @@ public class UninstallService extends Service {
             writer.write("wm density " + dpiValue + "\n");
         }
 
-        writer.write("rm -rf " + TEMP_DIR_NAME + "\n");
-        writer.write("rm -rf /data/data/*.tar.gz\n");
+        if (doUninstall) {
+            writer.write("rm -rf " + TEMP_DIR_NAME + "\n");
+            writer.write("rm -rf /data/data/*.tar.gz\n");
 
-        if (sourceDir.startsWith("/system")) {
-            disableApp(context);
-            writer.write("mount -o rw,remount /system\n");
-            writer.write("mount -o rw,remount /data\n");
-            writer.write("mount -o rw,remount /system/app/MigrateHelper\n");
-            writer.write("mount -o rw,remount /data/data/balti.migratehelper\n");
-            writer.write("rm -rf " + context.getApplicationInfo().dataDir + " " + sourceDir + "\n");
-        }
-        else {
-            writer.write("pm uninstall " + getPackageName() + "\n");
-        }
-
-        if (dpiValue > 0){
-            writer.write("reboot\nexit\n");
+            if (sourceDir.startsWith("/system")) {
+                disableApp();
+                writer.write("mount -o rw,remount /system\n");
+                writer.write("mount -o rw,remount /data\n");
+                writer.write("mount -o rw,remount /system/app/MigrateHelper\n");
+                writer.write("mount -o rw,remount /data/data/balti.migratehelper\n");
+                writer.write("rm -rf " + getApplicationInfo().dataDir + " " + sourceDir + "\n");
+            } else {
+                writer.write("pm uninstall " + getPackageName() + "\n");
+            }
         }
 
-        context.stopService(new Intent(context, StupidStartupService.class));
+        if (doReboot){
+            writer.write("reboot\n");
+        }
+
+        writer.write("exit\n");
+
+        stopService(new Intent(this, StupidStartupService.class));
 
         writer.flush();
         fullProcess.waitFor();
@@ -77,14 +92,14 @@ public class UninstallService extends Service {
 
     }
 
-    void disableApp(Context context){
+    void disableApp(){
 
-        SharedPreferences.Editor editor = context.getSharedPreferences("main", MODE_PRIVATE).edit();
+        SharedPreferences.Editor editor = getSharedPreferences("main", MODE_PRIVATE).edit();
         editor.putBoolean("isDisabled", true);
         editor.commit();
 
-        PackageManager packageManager = context.getPackageManager();
-        ComponentName componentName = new ComponentName(context, balti.migratehelper.MainActivity.class);
+        PackageManager packageManager = getPackageManager();
+        ComponentName componentName = new ComponentName(this, balti.migratehelper.MainActivity.class);
         packageManager.setComponentEnabledSetting(componentName, PackageManager.COMPONENT_ENABLED_STATE_DISABLED, PackageManager.DONT_KILL_APP);
     }
